create dxvk.conf

+++ Form1.cs
@@ -17,30 +17,55 @@ namespace dxvk_osd_customizer
             InitializeComponent();
         }
 
-            //if(File.Exists("dxvk.conf"))
-           // {
-            //    
+        //if(File.Exists("dxvk.conf"))
+        // {
+        //    
 
         private void buttonCreate_Click(object sender, System.EventArgs e)
         {
-            File.Create("dxvk.conf").Close();
-            using (StreamWriter sw = File.AppendText("dxvk.conf"))
-            sw.WriteLine("dxvk.enableAsync=true" +
-                    "\nRADV_PERFTEST = gpl" +
-                    "\ndxvk.gplAsyncCache = true");
-        }                
-                //{
-                
-                //}
-           /* }
+            string aString = "";
+            /*
+            foreach (Control c in this.Controls)
+            {
+                if (c is CheckBox)
+                {
+                    c.Controls.
+                }
+              */
+            if (checkBoxDevinfo.Checked)
+                aString = "\ndxvk.hud=devinfo";
+
+            /*
+            if (!File.Exists("dxvk.conf"))
+            {
+
+                File.Create("dxvk.conf").Close();
+                using (StreamWriter sw = File.AppendText("dxvk.conf"))
+                { sw.WriteLine("dxvk.enableAsync=true" +
+                            "\nRADV_PERFTEST = gpl" +
+                            "\ndxvk.gplAsyncCache = true"+aString); 
+                }
+             }
+                          
             else
             {
-            File.WriteAllText("dxvk.conf", string.Empty);
-            using (StreamWriter sw = File.AppendText("dxvk.conf")) ;
+                File.WriteAllText("dxvk.conf", string.Empty);
+                using (StreamWriter sw = File.AppendText("dxvk.conf"))
+                {
+                    sw.WriteLine("Write some text");
+                   
+                    }
+                }
+            */
+            File.Create("dxvk.conf").Close();
+            using (StreamWriter sw = File.AppendText("dxvk.conf"))
             {
-            sw.WriteLine("yoyo");
+                sw.WriteLine("dxvk.enableAsync=true" +
+                          "\nRADV_PERFTEST = gpl" +
+                          "\ndxvk.gplAsyncCache = true" + aString);
+
             }
-            }
-            }*/
-}
+
+        }
+    }
 }
